wip: refactor peps
