--- yastn/tn/fpeps/evolution/__init__.py
@@ -1,2 +1,0 @@
-""" imaginary time evolution routines. """
-from ._update import gates_homogeneous, evolution_step_, Gate_local, Gate_nn, Gates
